dcap: Use pin manager rather than the stager
PinManager largely provides the same functionality as the stager
component. Tigran tells me the stager was never completely finished.

Since we plan to make significant changes to PoolManager this patch
updates dcap to use PinManager. The patch moves the stager components.

Also fixed a bug in gsidcap and kerberosdcap: When used with space
manager these doors didn't redirect pool messages through space manager.
This would have caused access latency and retention policy of a space
token to be ignored when written to through dcap.

RELEASE NOTES: The dummy-stager services is obsolete and has been 
replaced by the pinmanager service. The -t option of the dccp command is 
obsolete and longer has any effect (that's the -stagetime option inside 
the protocol).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>