transfermanagers: Clean up CopyManager
To make it easier to modify PoolManager, I plan to move CopyManager to
use the Transfer facade. Before I do that there are some things that
needed to be cleaned up in CopyManager. That's done in this patch.

This also fixes a race condition in accessing a shared map.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3048