properties: better warning message for deprecated properties
Properties may be annotated as forbidden, obsolete, deprecated and 
not-for-service.  The deprecated annotation indicates that a property 
is currently supported but support for that property will disappear in 
the future.  If the user configures dCache with a deprecated property 
then dCache will generate a warning message, but continue to work.

The current warning message is limited to stating that support for that 
property will disappear (at some unspecified point in the future).  In 
many cases, a deprecated property is being replaced by a new property.  
To maintain backwards compatibility, the deprecated property will 
either be defined:

    new-property = some default value
    (deprecated)old-property = ${new-property}

or as:

    new-property = ${old-property}
    (deprecated)old-property = some default value

If the former style is used or the latter has only a single property 
with a simple reference then the name of the replacement property 
("new-property" in these examples) is knowable.

This patch adjusts the warning message if the replacement property is 
known.  If the new property isn't known then a slightly modified 
version of the warning message is printed:

    Property old-property: please review configuration; support
    for old-property will be removed in the future

If the replacement property is known then the warning message includes 
that information:

    Property old-property: use "new-property" instead; support
    for old-property will be removed in the future

Target: trunk
Patch: http://rb.dcache.org/r/2781/
Acked-by: Gerd Behrmann