webdav: fail with meaningful error if KeyStore or TrustStore is missing
IF webdav is enabled with the secure transport but either the KeyStore 
or TrustStore is absent then the log file contains a massive stack-trace 
(so definitely a bug) and it is difficult for the sysadmin to understand 
how to fix the problem.

This patch adds tests for the KeyStore and TrustStore files.  If they 
are missing, the domain doesn't start (as with current behaviour) but 
the error message explains what is wrong and suggests how to fix it.

Target: trunk
Patch: http://rb.dcache.org/r/3065
Acked-by: Gerd Behrmann