poolmanager: fix percentile cost calculation
The cost-cut for triggering pool2pool transfers on load may be 
specified as an absolute value or a percentile value.  If using a 
percentile value then cost is the CPU cost of the n-th pool of the m 
known pools, where n/m is the desired percentage value.  This cost-cut 
value is dynamic and expensive to calculate, so the value is cached and 
re-evaluated only when pool costs have changed sufficiently that the 
veracity of the cached value is in doubt.

Unfortunately, there are several bugs that result in the cache value 
being seldom invalidated: there are events that affect the recorded 
cost of a pool, but that currently cannot trigger invalidation of a 
cached cost-cut.  Instead, the cached cost-cut is re-evaluated only 
when a pool goes down or becoming available.

This patch fixes these problems.  After the patch, the periodic 
messages from pools (typically every 30 seconds) may invalidate the 
cached cost-cut value.  It also allows the pool-manager "magic" to 
trigger cache invalidation.

Target: trunk
Patch: http://rb.dcache.org/r/3049/
Acked-by: Gerd Bahrmann
Request: 1.9.5
Request: 1.9.9
Request: 1.9.10
Request: 1.9.11