webdav: Add size column to listing and use template engine for rendering
I got several requests to add the file size to the HTML rendering of
directory listing for WebDAV. This is done by this patch.

The patch introduces StringTemplate as a template engine. Thus we can
move the HTML part out of the Java code and into a template. Currently I
embed the template into the jar file, but I will probably make another
patch to move it into the share directory: This is a file that a site is
likely to customize.

The choice of StringTemplate was made after having a brief look at the
various libraries available. StringTemplate stands out by not being able
to do everything: It encourages (enforces) a clear separation between
view and model. Besides making the code cleaner, this also makes it
safer to let sites customize the view. StringTemplate is written by the
same guy who wrote ANTLR (a very popular parser generator). The library
is small and doesn't introduce strong dependencies into the code (ie. we
can always change it later if we find something better).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3082/