poolmanager: Fix message queue overflow handling in RequestContainerV5
RequestContainerV5 uses a ThreadPoolNG instance to execute state
machines for PoolMgrSelectReadPoolMsg requests. The state machine is
started on demand when the state machine needs to be executed. The
request handler maintains a boolean variable to track whether the state
machine isc currently active or not. If true the state machine will not
be started again.

ThreadPoolNG does not use a task queue (it uses a SynchronousQueue). If
configured with a thread limit the lack of a queue means that task
submission may failed with a RejectedExecutionException. Unfortunately
this happens after the request handler sets the aforementioned boolean
variable to true. After this point the handler believes the state
machine runs even though submitting it to the thread pool failed.
PoolManager never recovers from this situation and even an explicit 'rc
retry' attempt does not help. PoolManager needs to be restarted.

This patch adds a try-catch around the submission point and sets the
flag back to false if submission fails.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3080/