configuration: activate support for not-for-services annotation
A recent patch introduced the ability to annotate a property as 
not-for-services.  This is meant to indicate that the property has no 
effect when declared in a services scope; for example, setting one of 
the Java memory options cannot have an effect in any service scope.

Target: trunk
Patch: http://rb.dcache.org/r/3085/
Acked-by: Gerd Behrmann