gPlazma2: added VORoleMap Plugin
gPlazma2 is not able to do authorization and mapping based on vorolemap and authzdb
files, but to keep compatibility with gPlazma1 this is needed.
This patch adds a new plugin implementing authorization and mapping capabilities 
based on these files.

Patch: http://rb.dcache.org/r/2808/
Acked-by: Tigran
Target: trunk

RELEASE NOTES:
Added functionality to gPlazma2 to work with vorolemap and authzdb files.