CopyManager: Update CopyManager to use the Transfer facade
Pretty much a rewrite of most of CopyManager. As a consequence CopyManager now respects ACLs. It also submits to the pp and p2p queues to avoid transfer deadlocks.

It is rather easy to add "Active Transfers"-page and billing integration if we want to. This has however not been done yet.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3053/