cells: Fix missing timeout on message queue overflow
When sending a message a cell has two ways to receive a reply (actually
three, but let's forget about the third one here): It can use
sendAndWait or use send with a callback.

When CellNucleus receives a reply, it will check if somebody is waiting
for the reply with sendAndWait. If so, it will deliver the message right
away. If not, it will check if somebody registered a callback. If so, it
will enqueue the reply for delivery on the callback executor thread
pool.

If the callback executor thread pool has a capacity limit then enqueuing
may fail with RejectedExecutionException. This happens after the
callback has already been unregistered. The consequence is that the
reply will not be delivered, but the callback doesn't time out either.
The sender will thus wait for the reply forever.

This patch addresses this problem by registering the callback again
after enqueueing failed. This will ensure that the request eventually
times out.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3072/