transfermanagers: Generalize RemoteGsiftpTransferManager
This patch takes the gsiftp specific part out of the TransferManager and
puts it inside the Storage class. This allows us to add support for
HTTP. Although an HTTP transfer manager existed, SRM didn't actually
support using it. This patch adds the missing support.

The old HTTP transfer manager is removed.

The SRM client is updated to support http as a protocol.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3058/