config: add check-config command, fix line numbering
This patch adds support for the check-config command.  This is a 
/opt/d-cache/bin/dcache command that scans the dCache configuration 
and reports all the warnings or errors.  This is very useful as dCache 
will abort as soon as it detects an error in the configuration; any 
subsequent problems are only discoverable once the error has been 
fixed.  This results in a rather painful process of diagnosing and 
fixing problems in dCache configuration.

This patch also fixes the contextual information (filename and 
line-number) in configuration warning or error messages.  Currently, 
only the layout file messages contain a filename and line-number and, 
even then, the line-number is wrong.

With this patch, all configuration warning or error messages are 
printed with the name of the file and correct line number where the 
problem is found.

Target: trunk
Patch: http://rb.dcache.org/r/3087/
Acked-by: Gerd Behrmann