properties: add missing not-for-services and deprecated annotations
dCache configuration files may be annotated to provide metadata.  
This has two effects: it documents behaviour and allows dCache to 
provide warning or error messages against misconfigured sites.

The "not-for-services" annotation means that configuring this 
property has no effect inside a service (so we can warn a confused 
sys-admin if they attempt to configure the property in that scope).  
There is currently a lack of this annotation, which this patch adds.

The file dcache.properties contains a number of properties that are 
deprecated.  There is a preferred property name that we would like 
sites to use instead.  Sometimes this is described in the comments, 
but unfortunately not using the deprecated annotation.  This patch 
adds the deprecated annotation, where appropriate, and removes the 
redundant comment.

In addition, a few comments are tidied up.

Target: trunk
Patch: http://rb.dcache.org/r/3088/
Acked-by: Gerd Behrmann