all: Refactor MessageCallback
I ran into the problem that I need the Message reply for both successful
and failed operations. The current MessageCallback interface does not
provide this.

This patch refactors MessageCallback to provide access to the reply.
AbstractMessageCallback is introduced to simplift the task of
implementing the callback and to reduce the differences between the new
and old interface.

The noroute() and timeout() callbacks now take a CellPath as an
argument. This makes it easier to provide useful error messages.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3059/