gPlazma2: Configuration passing for VORoleMap and ARGUS plugins
Currently the Argus and VORoleMap Plugins expect their parameters to come in the form 'key1, value1, key2, value2, ...'.
gPlazma however gets the parameters as space separated, potentially quoted sequence of values.
This patch fixes the behaviour of the plugins in a way that now the parameters may be given in the form
'key1=value1 "key 2= value two" ... '
This allows passing arguments in an arbitrary order. From the list of arguments a map is created for a fixed set of 
expected keys (for each plugin). Missing keys, unexpected keys, and malformed key value pairs will raise an 
IllegalArgumentException.

Patch: http://rb.dcache.org/r/3098/
Acked-by: Paul
Target: trunk