configuration: fix discovery of synonyms
When generating warning messages about deprecated properties, dCache 
code looks for a unique synonym of the deprecated property.  This 
allows dCache to provide more meaningful warning messages: if dCache 
knows the name of the new property then it can include this in the 
warning.

Unfortunately, the synonym searching code searched only the 
properties in the current ConfigurationProperties and not the ones 
defined as defaults in parent Properties objects.  This breaks the 
synonym search when someone has configured a deprecated property in 
dcache.conf or the layout file.

This patch fixes the issue by searching all known properties for 
synonyms.

Target: trunk
Patch: http://rb.dcache.org/r/3101/
Acked-by: Karsten Schwank