configuration: tidy up output from check-config
The output from the check-config (/opt/d-cache/bin/dcache) command is 
rather ugly.  In particular, it doesn't respect that lines may have a 
finite width due to terminal size constraints.

This patch formats the output from check-config so each line is no 
longer that 75 characters in length.  Lines that are longer are 
wrapped.  The patch also provides support for indenting lines.  The 
indentation is controlled by a regular expression: when the RE matches, 
the line up to that point is considered the prefix and subsequent lines 
are indented to match.

Target: trunk
Patch: http://rb.dcache.org/r/3113
Acked-by: Gerd Behrmann