migration script: Generate `hostname`.conf rather than imported.conf
Best practice recommendation is to use the host name as the name of the
layout file. Our migration script violated this recommendation by using
the name 'imported.conf'; this is a name that we can almost guarantee
will have to be changed after migration and is thus a rather bad name.

This patch updates the migration script to use the host name instead.
The patch also refactors the migration script to avoid repeating the
same file names over and over.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3112/