gPlazma2: Fix parameter handling by VORoleMap Mapping Plugin
The current implementation of GPlazmaVoRoleMap.map maps principals found in "authorizedPrincipals" and adds the mapped Principals to "principals".
This is not correct, since in GPlazma the authentication step fills a set of principals and passes them on to the mapping step as "principals". Following steps then use the mapped principals from the "authorizedPrincipals" set.
This patch fixes the plugins behaviour. 

Patch:http://rb.dcache.org/r/3103
Acked-by:Paul
Target:trunk