pool: fix way too short timeout for http transfer
The new http mover, introduced with 1.9.11, includes the 
httpMoverClientIdleTimeout property, which is described as "Period in 
seconds after which a client will be disconnected if the connection 
is idle (not reading or writing)".  The default value is 300 (seconds 
== five minutes).

Unfortunately, the units for this property in 1.9.11, 1.9.12 and 
trunk is actually milliseconds.  If the client or server is idle for 
300 ms then the pool will disconnect.  There is a real risk that a 
transfer may stall resulting in a long enough for the pool to kill 
the transfer.  The longer the file, the more likely that there will 
be a pause of 300 ms or longer.

The patch changes the timeout to be in seconds.

Target: trunk
Request: 1.9.12
Request: 1.9.11