migration: handle RPM removing or renaming chimera-config.xml
The migration script for converting dCache 1.9.5 configuration into 
the format for 1.9.12 assumes the existence of the file 
(/opt/d-cache/) config/chimera-config.xml.  This file is read to 
create the appropriate configuration for access to the Chimera 
database.

Unfortunately, when upgrading dCache v1.9.5 to v1.9.12, RPM will 
either delete the chimera-config.xml file (if it has not been edited) 
or rename it to chimera-config.xml.rpmsave if it has been edited.  
Since the chimera-config.xml file no longer exists, the migration 
process fails.

This patch fixes this problem.  It tries to find a suitable 
"chimera-config.xml"-like file from three possible locations.  These 
are (in order)

    /opt/d-cache/config/chimera-config.xml
    /opt/d-cache/config/chimera-config.xml.rpmsave
    /opt/d-cache/share/migration/chimera-config-from-195.xml

The first file found from the list (in list order) is used for the 
migration.  If no file is found then the migration will fail with an 
error message.

The file chimera-config-from-195.xml is a copy of the 
chimera-config.xml file from the 1.9.5 branch's HEAD (although the 
commented out part has been removed).

Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9