migration: fix use of tail
Solaris and Linux have different tail commands.  Fixing the migration 
script to work with Solaris broke the script for Linux machines.  
From some experimentation, it seems impossible to have a single 
command that works on both platforms when selecting lines counting 
from the top of a file.

This patch introduces a function that chooses an invocation of tail 
for Linux or Solaris depending on the running platform.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3132/
Acked-by: Gerd Behrmann