httpd: Introduce cleaned up version of HttpServiceCell
The cells package provides the HttpServiceCell with custom HTTP support.
The httpd service is build on this cell.

This patch introduces a cleaned up copy of HttpServiceCell. The new cell
uses Jetty for HTTP handling and cleans up a few things, but is
otherwise identical to the old cell. I deliberately tried to reduce how
much is changed.

Eventually we may move to using more of Jetty's features.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>