jms: Rename tunnel options to explicitly refer to amq
JMS is an API, not a wire protocol and not a broker implementation. Thus
it doesn't make much sense to refer to 'jms' as a broker in our
configuration.

This patch renames the options to explicitly refer to 'amq' as the
broker. This matches the configuration parameters, which all start with
amq.

'jms' is renamed to 'amq'. 'hybrid' is renamed to 'cells+amq-embedded'.
The old names are still supported but generate log messages when used.

The new scheme 'amq-embedded' is added to support an embedded broker
without cells support.

The patch also fixes a couple of cases in which we relied on deprecated
properties.

This patch patch prepares for introducing an alternative JMS broker:
HornetQ.

Since the patch also fixes references to deprecated parameter names I
suggest it is merged into 1.9.12.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3147/