doors: Use a session ID compatible with InfoMessage.getTransaction
The Transfer class generates a transaction id for each transfer. This ID
is logged in the billing files.

In dCache 1.9.5 the transaction ID format is defined by
InfoMessage.getTransaction. The format includes the cell name and domain
name of the door. This had the nice side effect that one could easily
link the billing entries generated by the door and by the pool.

The Transfer class uses the CDC session key as a base for generating a
transaction ID for a transfer. The CDC session key only included the
domain name thus making it harder to match the door and pool billing
entries.

Even though the two entries in 1.9.5 only match by accident, I still
consider this change a regression. This patch addresses that regression
by letting doors generate a session key compatible with the format in
InfoMessage.getTransaction.

Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10
Acked-by: Tanja Baranova <tatjana.baranova@desy.de>
Patch: http://rb.dcache.org/r/3149/