doors: add listen option
The LoginManager class has the option to listen (bind) to a particular interface.  This patch 
makes this option available to sys-admins via the listen property.  If the value of this 
property is the empty string then the door will listen on all interfaces; otherwise, the value 
is the IP address of the interface the door should listen to.

The patch also allows the SRM to bind to a particular interface using the same property.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3144/
Acked-by: Tigran Mkrtchyan