cells: Introduce central cell name service for JMS brokers
A JMS broker presents the challenge to dCache that queues are decoupled
from the receivers. That's of course one of the great advantages,
however it also means that there is no such thing as
NoRouteToCellException; a message can be delivered to a queue even when
there are no receivers.

This is mainly an issue for interactive users when the user mistypes a
cell name, however that's still a usability regression compared to
classic cells messaging.

Well known cell resolution in dCache relies on a broadcast based address
resolution protocol (ARP). Broadcasting a message to thousands of
domains may not be the most scalable thing to do (even though we could
possibly optimize that by using message selectors).

This patch tries to address both issues by introducing a central cell
name service (CNS). Any domain registers its well known cell with the
service and other cells can ask the service to learn which domain
contains a particular well known cell.

The registration with the CNS is done over a JMS topic (ie
publish-subscribe) and thus any number of CNSes can be added. Lookups
are performed by sending a query on a queue from which all CNS instances
receive messages; since it is a queue only one CNS instance will receive
a particular query.

JMSTunnel is updated to register with CNS and to submit lookups to CNS.
Like before replies are cached in JMSTunnel. Topology cell is updated to
embed the CNS registry. Thus it will automatically learn about all
domains without having to query another service.


Compatibility with the old ARP system is not maintained: A synchronous
update is required if JMS is used.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3156/