srm: Add per request type database parameters
There are several database related parameters beyond the connection
parameters in the SRM. These are parameters that control whether to
store jobs in a DB, whether to store state transitions, for how long to
maintain requests in the database, etc. Currently these parameters apply
to any request using the scheduler infrastructure.

When we added the asynchronous ls feature, ls request moved to the
scheduler infrastructure. Initially this was optional, however in recent
releases this code path is used no matter whether asynchronous ls is
actually enabled or not (that because we can now support a mode in which
we dynamically switch between the sychronous and asynchronous mode). A
consequence of that is that all ls requests are now stored in the DB
too.

This change is of course great for monitoring (you can now see who is
issuing all those ls requests and what they are listing), but also means
that a lot more information was stored in the DB. NDGF recently
discovered that some 20 GB of our database was used just for storing ls
requests. We feel that this is maybe a bit excessive (keeping in mind
that the complete Chimera DB is smaller than that).

This patch refines the SRM so that storage to the DB, rentention time,
etc, can be controlled per request type. Eg. one could limit ls requests
to 1 day or disable storage of those requests in the database.

Since the huge increase in data in the database is to be considered a
performance regression, I recommend that this is merged into 1.9.12.

RELEASE NOTES: Several parameters got renamed in the process and the old 
names have been marked as deprecated. The migration script will be 
updated in a separate patch.

Target: trunk
Request: 1.9.12