pool: fix waitForFiles
The UniversalSpringCell supports the waitForFiles property that allows 
any cell that uses UniversalSpringCell to block until a supplied list of 
files are present in the filesystem.  Although this property will work 
for all such cells, it is currently used only for pools, where dCache 
may be configured to block starting up the pool until a file within the 
pool is visible.  This is to prevent dCache from auto-creating the pool 
infrastructure and starting up if a remote filesystem has failed to 
mount (or the filesystem is mounted but not yet visible).

The current behaviour is broken: the constructor of various components 
will auto-create their required files and directories within their 
constructors.  Since UniversalSpringCell eagerly creates instances of 
these classes before testing the presence of waitForFiles-supplied 
files, the waitForFiles fails to operate correctly.

This patch fixes the problem by moving when the waitForFiles property is 
considered.  By moving the point to before when the UniversalSpringCell 
context is established, the block now happens before the auto-creation 
of the cell beans.

Target: trunk
Request: 1.9.12
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6224
Acked-by: Gerd Behrmann
Acked-by: Tanja Baranova