doors: Refine semantics of PoolMgrSelectReadPoolMsg
This patch refines the semantics of PoolMgrSelectReadPoolMsg. In
particular the behaviour in case of failures is defined.

Doors and services are updated to comply with the new semantics.

This patch prepares for a patch which updates PoolManager to no longer
retry read requests internally. Instead PoolManager will ask the
requestor (the doors) to retry. This patch adds this retry logic to the
doors.

Since PoolManager needs to maintain state between retries,
PoolMgrSelectReadPoolMsg has to ability to include previous replies from
PoolManager. Currently there is no useful state information in the
message, however such information will be added when PoolManager is
updated.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3096/