PoolManager: Use location in FileAttributes and move retry loop to doors
PoolManager currently talks to PnfsManager to learn about file
locations. This adds an extra round trip to PnfsManager, which increases
the latency of PoolManager. PoolManager also uses a blocking call for
talking to PnfsManager, which means that a thread was occupied.

This patch removes the lookup and instead relies on the locations being
present in the FileAttributes provided by the door. The door can fetch
this information together with storage info, thus avoiding the extra
round trip to PnfsManager.

One complication is that PoolManager at several points refreshes the
file location and in some cases storage info from PnfsManager. It needs
to do this after p2p and stage, but also when retrying in case of errors
(eg we are waiting for a file to be flushed to tape). Rather than
letting both doors and PoolManager contain logic to fetch this
information, this patch removes the retry loop from PoolManager. Instead
PoolManager relies on doors to retry if they wish to do so. This also
avoids nested retry loops in which both doors and PoolManager retry.

PoolManager uses the newly added OutOfDateCacheException to request an
immediate retry when PoolManager knows that file meta data has changed.
PoolManager embeds the retry counter and last stage host in the
PoolMgrSelectReadPoolMsg to ensure that this information survives a
retry. PoolManager still implements the suspend logic, although it
relies more on the door with respect to how quickly to retry after a
failure and whether to retry at all.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3165/