gplazma: Decouple gPlazma from SessionAttributes
A reply to a login contains a set of attributes associated with the
login/session. In gPlazma these are represented as implementations of
SessionAttribute. In dCache a similar concept is represented as
implementations of LoginAttributes.

Since gPlazma strives to be independent of dCache and dCache strives to
be independent of gPlazma, the above mismatch requires that attributes
are converted in the LoginStrategy interfacing with gPlazma. This in
turn means that a major design goal is lost: The ability for plugins to
add new attributes unknown to gPlazma and the plumbing in dCache (eg a
plugin that adds attributes known to some new custom door). Plugins are
limitted to the attributes known to the conversion code.


To remove this flaw this patch removes the SessionAttribute interface.
Instead gPlazma operates on a set of objects (no specific interface).
Specific gPlazma plugins support the LoginAttributes used by dCache.
Those are in turn moved to dCache commons.

Thus the gPlazma framework is independent of dCache and even dCache
commons. dCache is independent of gPlazma. Specific gPlazma plugins
depend on dCache commons which contains dCache specific principals and
attributes.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3221/