scripts: don't start if non-migrated config is detected
A sys-admin may attempt to upgrade to 1.9.12 and start dCache.  
Although this will "work", it will not give the expected behaviour 
and the sys-admin may perceive this as a failure and either downgrade 
or issue a support ticket.

This patch aims to alert the sys-admin to their failure to read the 
release notes.  It uses the presence and absence of key configuration 
files as a heuristic for detecting whether the site-admin has 
upgraded their dCache instance without converting their 
configuration.  If this situation is detected then the 
(/opt/d-cache/bin/)dcache script will abort almost all operations 
with a suitable error message and a non-zero return code.  The script 
was based on the idea from Gerd.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3225/
Acked-by: Gerd Behrmann