admin: move default location of ssh keys to /opt/d-cache/etc
In previous versions of dCache, the ssh keys are stored in the 
/opt/d-cache/config directory.  A more standards-compliant location for 
these files is /opt/d-cache/etc.

This patch changes the default location for the ssh-keys to be 
/opt/d-cache/etc.

The patch also changes the property name from keyBase to 
dcache.paths.ssh-keys.  The new property is backwards compatible with 
dCache instances that configuration the old property (keyBase).  The 
migration script is updated to migrate a site's configuration that 
configures keyBase to use the new property name.

The migration script is updated so it will copy the ssh key-related 
files into /opt/d-cache/etc if they are found in the standard locations 
(in /opt/d-cache/config) and the source files are renamed to have the 
'.pre-197' suffix, in keeping with other migrated files.

Target: trunk
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6226