lm: Fix setup file creation problem for non-root installation
We do not ship with an lm.config file. LocationManager however tries to
create an empty configuration file during startup. This fails when
dCache runs as a user without write permission to the dCache etc
directory.

This patch fixes the issue by not considering the failure to create the
setup file a fatal error.

Notice that the old code falsely used the boolean return value of
createNewFile as an indicator of whether the file failed to be created.

Target: trunk
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6221
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3237/