admin: use dcache.paths.ssh-keys property rather than keyBase property
In 1.9.12, the keyBase property is marked deprecated.  When migrating 
a dCache configuration from 1.9.5 to 1.9.12, a property assignment 
statement that affects keyBase is automatically updated to the new 
property: dcache.paths.ssh-keys.

In the admin.properties file, dcache.paths.ssh-keys is defined in 
terms of the deprecated property's value (keyBase).  If a site-admin 
has modified keyBase then the migration process will modify this 
statement to assign a value to the dcache.paths.ssh-keys property.  
The value of keyBase will revert to the default value.  Any references 
to keyBase will provide the "wrong" value.

Unfortunately, the admin.batch file contains a reference to keyBase.  
This will break a site if they upgrade and migrate their 
configuration.

This patch fixes this by replacing the reference to the deprecated 
property with a reference to the new property.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3249/
Acked-by: Gerd Berhmann