HSM: throw a meaningful error if max-active set negative
This patch follows the idea from an old patch in RT:

	http://rt.dcache.org/Ticket/Display.html?id=3522

The idea is to provide a meaningful error message if the sys-admin 
attempts to set a negative limit on the number of concurrent restores.  
As the bulk of the scheduling happens elsewhere with current trunk, a 
separate patch will address this problem for IO queues.

Target: trunk
Patch: http://rb.dcache.org/r/3252/
Acked-by: Gerd Behrmann