packaging: Make DEB specific FHS packaging generic
Refactors the logic for restructuring the directory structure to an FHS
compliant structure. Rather than performing the restructuring in the DEB
packaging we now do it in an ant task.

The new ant target install-server-fhs produces an FHS compliant
installation structure. Ant filtersets are used to replace tokens in
various script and properties files with the proper values.

To simplify the process I moved several files from modules/ to skel/

The classic packaging structures should be unaffected.

Notice that the approach taken is still a workaround: The bulk of the
build targets the old directory structure. FHS packaging then transforms
that structure to an FHS compliant structure. We should eventually
refactor the build itself to use a flexible directory structure.

Target: trunk
Request: 1.9.12
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/3254/