cells: fix updating CNS if a domain is stops running a well-known cell
The new CellNameService (CNS) provides a lookup service for well-known 
cells.  It uses soft-state where the domains push a list of their 
running well-known cells on a regular basis.  Using the CNS should 
help prevent long delays when a user attempts to "cd" into a 
non-existent well-known cell.

There are several places where activity triggers removing the cached 
information about a domain.  Unfortunately, there is a bug where 
removing the domain doesn't remove all the cells in that domain.  The 
effect is that a domain is not running a well-known cell that it was 
previously running and that cell is not running elsewhere then CNS 
will return the old domain when queried about this cell.  This 
probably affects doors most.

The patch fixes the problem (a simple typo).

It also introduces some unit-tests (~76% coverage) for the 
CellNameServiceRegistry class.  One of the tests 
(testDomainUpdatesWithOneLessCell) fails for current trunk/1.9.12 but 
passes after the fix.

Kudos: FindBugs
Target: trunk
Request: 1.9.12 
Patch: http://rb.dcache.org/r/3258/
Acked-by: Gerd Behrmann