files: migrate admin interface authz files to new location
dCache 1.9.12 and trunk uses three subdirectories of 
/opt/d-cache/etc/users for authz information ('acls', 'meta' and 
'relations').  Previous versions of dCache used three subdirectories of 
/opt/d-cache/config/users directory (with the same names) for this 
purpose.

A previous patch changed the location of this authz information from 
config/users to etc/users.  However, the patch was incomplete because 
it failed to update the migration script and failed to note (for 
release notes) that manual intervention is needed for sites that are 
either not using the migration script or that have already updated to 
1.9.12-1.  The patch also failed to mention (for the release notes) 
that the location of the admin interface's ssh-keys has also moved.

This patch adds the missing functionality to the migration script.  The 
contents of the config/users directory is copied to the equivalent 
location within the etc/users directory and the config/users directory 
is renamed to config/users.pre-197.

FOR RELEASE NOTES:

The location of files and directories needed by the SSH admin interface 
have moved.  The location of the four SSH key-pair files, needed by the 
admin interface, have moved from /opt/d-cache/config to 
/opt/d-cache/etc.  The location of authorisation information previously 
located in /opt/d-cache/config/users is now located in 
/opt/d-cache/etc/users.

If you are upgrading from 1.9.5, the migration script 
/opt/d-cache/libexec/migrate-from-1.9.5.sh will manage this change and 
no manual intervention should be needed.

If you have already upgraded to 1.9.12-1 or decide not to use the 
migrate-from-1.9.5.sh script then you must ensure that the correct 
information is available in /opt/d-cache/etc.  One way to achieve this 
is simply to move the four SSH-related files ('host_key', 
'host_key.pub', 'server_key' and 'server_key.pub') from the 
/opt/d-cache/config directory to /opt/d-cache/etc and to recursively 
copy the contents of /opt/d-cache/config/users into the 
/opt/d-cache/etc/users directory.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3264
Acked-by: Gerd Behrmann