scripts: fix location of node_config in test for non-migrated dCache
A recent patch introduced a test to check if a site-admin is 
mistakenly attempting to start dCache without first migrating the 
site's configuration.  This uses the presence of two files as a 
heuristic that indicates that the dCache instance’s configuration 
hasn't been upgraded.

Unfortunately, the script looks for one of the two files (node_config) 
in /opt/d-cache/config.  Since the file is located in 
/opt/d-cache/etc, the heuristic will always fail, and the check offers 
no protection against non-migrated sites.

This patch fixes the problem by looking for node_config in the correct 
location.

Target: trunk
Request: 1.9.12
Patch: http://rb.dcache.org/r/3265/
Acked-by: Gerd Behrmann